requirements.txt
polib
requests
web.py
pycairo
drawsvg
